Add files via upload
This is the version submitted to Coursera for grading. It did not pass, giving the grader output of  "Dataframe does not contain features field, available fields are: columns data index". Moreover,  I tested zipping/unzip and encode/decode on my machine. It works fine on my machine (used Spyder notebook)